fix: cI and lint configuration for go.mod limits
- Add search_memory_test.go, lint_regression_test.go
- Update document.go and 31 others

+++ search/searcher/docstub_test.go
@@ -115,8 +115,9 @@ type FakeField struct {
 }
 
 func NewFakeGeoField(name string, lon, lat float64) *FakeField {
-	mHash := geo.MortonHash(lon, lat)
-	prefixCoded := numeric.MustNewPrefixCodedInt64(int64(mHash), 0)
+	//nolint:gosec // G115: preserve all Morton hash bits in the signed prefix-coding API.
+	mHash := int64(geo.MortonHash(lon, lat))
+	prefixCoded := numeric.MustNewPrefixCodedInt64(mHash, 0)
 
 	rv := &FakeField{
 		N:  name,
@@ -125,7 +126,7 @@ func NewFakeGeoField(name string, lon, lat float64) *FakeField {
 		DV: true,
 	}
 
-	terms := addShiftTokens([]numeric.PrefixCoded{prefixCoded}, int64(mHash), 9)
+	terms := addShiftTokens([]numeric.PrefixCoded{prefixCoded}, mHash, 9)
 	for _, term := range terms {
 		ft := &FakeTerm{
 			T: string(term),
